requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.1
